Implemented new web server
